@@ -1,5 +1,5 @@
 <?php
-namespace samdark\sitemap;
+namespace SamDark\Sitemap;
 
 use XMLWriter;
 
@@ -36,7 +36,7 @@ class Index
     /**
      * Creates new file
      */
-    private function createNewFile()
+    private function createNewFile(): void
     {
         $this->writer = new XMLWriter();
         $this->writer->openMemory();
@@ -53,14 +53,8 @@ class Index
      * @param integer $lastModified unix timestamp of sitemap modification time
      * @throws \InvalidArgumentException
      */
-    public function addSitemap($location, $lastModified = null)
+    public function addSitemap($location, $lastModified = null): void
     {
-        if (false === filter_var($location, FILTER_VALIDATE_URL)) {
-            throw new \InvalidArgumentException(
-                "The location must be a valid URL. You have specified: {$location}."
-            );
-        }
-
         if ($this->writer === null) {
             $this->createNewFile();
         }
@@ -77,7 +71,7 @@ class Index
     /**
      * @return string index file path
      */
-    public function getFilePath()
+    public function getFilePath(): string
     {
         return $this->filePath;
     }
@@ -85,7 +79,7 @@ class Index
     /**
      * Finishes writing
      */
-    public function write()
+    public function write(): void
     {
         if ($this->writer instanceof XMLWriter) {
             $this->writer->endElement();
@@ -103,9 +97,9 @@ class Index
      * @param bool $value
      * @throws \RuntimeException when trying to enable gzip while zlib is not available
      */
-    public function setUseGzip($value)
+    public function setUseGzip($value): void
     {
-        if ($value && !extension_loaded('zlib')) {
+        if ($value && !\extension_loaded('zlib')) {
             throw new \RuntimeException('Zlib extension must be installed to gzip the sitemap.');
         }
         $this->useGzip = $value;
